requirements_AFC.txt
accelerate==1.2.1
ace_tools==0.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.10
aiohttp-cors==0.8.1
aiosignal==1.3.2
airportsdata==20250523
altair==5.5.0
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.8.0
astor==0.8.1
asttokens==3.0.0
async-timeout==4.0.3
asyncio==3.4.3
attrs==24.3.0
beautifulsoup4==4.12.3
blake3==1.0.5
bs4==0.0.2
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
cohere==5.13.6
colorful==0.5.6
comm==0.2.2
compressed-tensors==0.8.1
contourpy==1.3.0
cssselect==1.2.0
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
depyf==0.18.0
dill==0.4.0
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
einops==0.8.1
emoji==2.14.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.6
fastavro==1.10.0
feedfinder2==0.0.4
feedparser==6.0.11
ffmpy==0.5.0
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
gguf==0.10.0
google-api-core==2.24.2
google-auth==2.37.0
google-genai==0.4.0
google-search-results==2.4.2
googleapis-common-protos==1.70.0
googlesearch-python==1.2.5
gradio==3.36.1
gradio_client==0.2.9
greenlet==3.1.1
groq==0.15.0
grpcio==1.71.0
h11==0.14.0
hf-xet==1.1.2
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.32.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
interegular==0.3.3
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.5
jedi==0.19.2
jieba3k==0.35.1
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
langchain-core==0.3.72
langchain-text-splitters==0.3.5
langid==1.1.6
langsmith==0.4.10
lark==1.2.2
lightning-utilities==0.12.0
linkify-it-py==2.0.3
lm-format-enforcer==0.10.9
lxml==5.3.0
lxml_html_clean==0.2.2
markdown-it-py==2.2.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistral_common==1.5.6
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
narwhals==1.22.0
nest-asyncio==1.6.0
networkx==3.2.1
newsapi-python==0.2.7
newspaper3k==0.2.8
ninja==1.11.1.4
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.55.3
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.10.14
outlines==0.1.11
outlines_core==0.1.26
packaging==25.0
pandas==2.2.3
parameterized==0.9.0
parso==0.8.4
partial-json-parser==0.2.1.1.post5
peft==0.14.0
pexpect==4.9.0
pillow==11.1.0
pip-system-certs==4.0
platformdirs==4.3.6
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.0
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.26.1
protobuf==6.30.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py-spy==0.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytorch-lightning==2.5.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
ray==2.46.0
referencing==0.35.1
regex==2024.11.6
replicate==1.0.4
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==3.3.1
sentencepiece==0.2.0
serpapi==0.1.5
sgmllib3k==1.0.0
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tinysegmenter==0.3
tldextract==5.1.3
tokenizers==0.21.0
torch==2.5.1
torchmetrics==1.6.2
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.3
triton==3.1.0
types-requests==2.32.0.20241016
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==2.3.0
uvicorn==0.31.0
uvloop==0.21.0
virtualenv==20.31.2
vllm==0.6.6.post1
watchfiles==1.0.5
wcwidth==0.2.13
websockets==14.1
widgetsnbextension==4.0.13
wrapt==1.17.2
xformers==0.0.28.post3
xgrammar==0.1.19
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
